removed default behavior of creating sdcard folder, respect sdcard tag file, copy sdcard detach.txt when module detach.txt is missing

+++ src/service.sh
@@ -6,17 +6,23 @@ sleep 1
 MODDIR="${0%/*}"
 
 sdcard_folder="/sdcard/DynamicDetachX"
-[ -f "$MODDIR/detach.txt" ] && [ -f "$sdcard_folder/detach.txt" ] && {
-    if [ -f "$sdcard_folder/replace" ];then 
+[ -f "$sdcard_folder/detach.txt" ] && {
+    if [ -f "$MODDIR/detach.txt" ]; then
+        [ -f "$sdcard_folder/enable" ] && {
+            if [ -f "$sdcard_folder/replace" ];then 
+                cp -rf "$sdcard_folder/detach.txt" "$MODDIR/detach.txt"
+            elif [ -f "$sdcard_folder/update" ];then
+                sort -u "$MODDIR/detach.txt" "$sdcard_folder/detach.txt" > "$MODDIR/tmp_detach.txt"
+                cp -rf "$MODDIR/tmp_detach.txt" "$MODDIR/detach.txt"
+                rm -rf "$MODDIR/tmp_detach.txt"
+            fi
+            [ -f "$sdcard_folder/mirror" ] && {
+                cp -rf "$MODDIR/detach.txt" "$sdcard_folder/detach.txt"
+            }
+        }
+    else
         cp -rf "$sdcard_folder/detach.txt" "$MODDIR/detach.txt"
-    elif [ -f "$sdcard_folder/update" ];then
-        sort -u "$MODDIR/detach.txt" "$sdcard_folder/detach.txt" > "$MODDIR/tmp_detach.txt"
-        cp -rf "$MODDIR/tmp_detach.txt" "$MODDIR/detach.txt"
-        rm -rf "$MODDIR/tmp_detach.txt"
     fi
-    [ -f "$sdcard_folder/mirror" ] && {
-        cp -rf "$MODDIR/detach.txt" "$sdcard_folder/detach.txt"
-    }
     
     # cleanup
     rm -rf "$sdcard_folder/enable"
@@ -26,4 +32,3 @@ sdcard_folder="/sdcard/DynamicDetachX"
     chown root:root "$MODDIR/detach.txt"
     chmod 0755      "$MODDIR/detach.txt"
 }
-[ ! -d "$sdcard_folder" ] && mkdir -p "$sdcard_folder"
